There are now hardcoded default Underscore templates for each type and the core. You can override that with Underscore templates in the page. If they are present in the DOM, they get used. Their script blocks must have IDs like:
a2-template-FOO

Where FOO could be 'contentWrapper' for the core's wrapper that it puts around each content object, or 'contentType-TYPENAME' for the edit view template for each type.

Hardcoded default Underscore templates make it easier to test and develop the browser side code. They also open the door to alternative backends.

This doesn't mean there is raw markup in the body of functions. The default templates are in distinct properties for the type and the core, not sprinkled in the middle of things.