* Drag and drop reordering of contents within a single area (not across areas yet) * Deleting contents * Adding contents
We now have a fully functional contentArea as far as the browser side goes... if all you want is a series of plain textarea contents, that is. (: But all the code specific to them is in a separate .js file. You can add a second type without hacking anything.

My next step is to turn our sf2 fileuploader bundle into the basis for a slideshow/files slot. That will require that some server side excitement go down.